fix(reference): use integer score weights

+++ internal/bankverticals/vendor_acceptance_forms.go
@@ -106,7 +106,7 @@ func responsePolicyAcceptanceFormInput(programID, legalEntityID string) monitori
 }
 
 func responsePolicyAcceptanceScoreProfile() *formcontract.ScoreProfile {
-	weightedYes := func(id, fieldID, label string, weight float64) formcontract.ScoreContribution {
+	weightedYes := func(id, fieldID, label string, weight int) formcontract.ScoreContribution {
 		return formcontract.ScoreContribution{ID: id, Label: label, Weight: weight, Required: true,
 			Predicate:   formcontract.Predicate{FieldID: fieldID, Operator: formcontract.PredicateEquals, Values: []string{"Yes"}},
 			MatchPoints: 100, NonMatchPoints: 0, Missing: formcontract.MissingIndeterminate}
